# Add command to compare configurations