Adds the `compare-configurations` command that takes two configurations from the statistics files and show any significant changes between the two. For example:

```
Summary:
========
- Status Changes: 14 (positive: 6)
- Runtime Changes: 18 (positive: 10)
- Objective Changes: 1 (posxive: 1)


Status Changes:
===============

Negative Status Changes:
------------------------
OPTIMAL_SOLUTION -> SATISFIED:
  - data/mznc2020/skill-allocation/skill_allocation_only.mzn data/mznc2020/skill-allocation/skill_allocation_mzn_1m_1.dzn
SATISFIED -> UNKNOWN:
  - data/mznc2020/stable-goods/stable-goods-solution.mzn data/mznc2020/stable-goods/s-d6.dzn
  - data/mznc2019/median-string/median_string_dp.mzn data/mznc2019/median-string/p2_10_8-3.dzn
  - data/mznc2019/rotating-workforce/rotating-workforce.mzn data/mznc2019/rotating-workforce/Example789.dzn
  - data/mznc2020/skill-allocation/skill_allocation_only.mzn data/mznc2020/skill-allocation/skill_allocation_mzn_2m_3.dzn
OPTIMAL_SOLUTION -> UNKNOWN:
  - data/mznc2020/skill-allocation/skill_allocation_only.mzn data/mznc2020/skill-allocation/skill_allocation_mzn_2m_1.dzn
  - data/mznc2020/sdn-chain/model.mzn data/mznc2020/sdn-chain/d10n780-2.dzn
  - data/mznc2020/sdn-chain/model.mzn data/mznc2020/sdn-chain/d10n780-1.dzn

Positive Status Changes:
------------------------
UNKNOWN -> OPTIMAL_SOLUTION:
  - data/mznc2019/hrc/hrc.mzn data/mznc2019/hrc/exp1-1-5110.dzn
  - data/mznc2020/hoist-benchmark-for-minizinc/hoist-benchmark.mzn data/mznc2020/hoist-benchmark-for-minizinc/PU_2_8_1.dzn
ERROR -> UNKNOWN:
  - data/mznc2020/pentominoes/pentominoes-int.mzn data/mznc2020/pentominoes/04.dzn
SATISFIED -> OPTIMAL_SOLUTION:
  - data/mznc2020/racp/racp.mzn data/mznc2020/racp/j30_26_2_1.0.dzn
UNKNOWN -> SATISFIED:
  - data/mznc2020/cable_tree_wiring/ctw.mzn data/mznc2020/cable_tree_wiring/R046.dzn
  - data/mznc2020/tower_challenge/tower.mzn data/mznc2020/tower_challenge/tower_070_070_15_070-04.dzn

Timing Changes (>±10.0%):
=========================
- (205.2%: 14.2s -> 43.4s) data/mznc2019/zephyrus/zephyrus.mzn data/mznc2019/zephyrus/12__10__8__3.dzn
- (127.1%: 65.3s -> 148.4s) data/mznc2019/amaze/amaze3.mzn data/mznc2019/amaze/2012-03-29.dzn
- (112.4%: 7.5s -> 15.8s) data/mznc2019/zephyrus/zephyrus.mzn data/mznc2019/zephyrus/12__8__6__3.dzn
- (98.4%: 431.6s -> 856.3s) data/mznc2020/sdn-chain/model.mzn data/mznc2020/sdn-chain/d24n600-3.dzn
- (57.3%: 105.6s -> 166.1s) data/mznc2019/rotating-workforce/rotating-workforce.mzn data/mznc2019/rotating-workforce/Example1337.dzn
- (28.6%: 23.7s -> 30.5s) data/mznc2019/zephyrus/zephyrus.mzn data/mznc2019/zephyrus/14__6__6__3.dzn
- (23.7%: 30.0s -> 37.1s) data/mznc2019/zephyrus/zephyrus.mzn data/mznc2019/zephyrus/14__10__6__3.dzn
- (14.9%: 163.0s -> 187.3s) data/mznc2019/triangular/triangular.mzn data/mznc2019/triangular/n10.dzn
- (-13.8%: 18.5s -> 15.9s) data/mznc2019/zephyrus/zephyrus.mzn data/mznc2019/zephyrus/12__6__6__3.dzn
- (-17.5%: 1.6s -> 1.3s) data/mznc2019/ptv/pax_model.mzn data/mznc2019/ptv/pax_period_20_2b.dzn
- (-25.0%: 1.8s -> 1.4s) data/mznc2019/ptv/pax_model.mzn data/mznc2019/ptv/pax_period_20_5b.dzn
- (-25.3%: 97.5s -> 72.9s) data/mznc2019/hrc/hrc.mzn data/mznc2019/hrc/exp1-1-5460.dzn
- (-25.7%: 1.8s -> 1.4s) data/mznc2019/ptv/pax_model.mzn data/mznc2019/ptv/pax_period_20_3b.dzn
- (-53.9%: 946.1s -> 436.5s) data/mznc2020/hoist-benchmark-for-minizinc/hoist-benchmark.mzn data/mznc2020/hoist-benchmark-for-minizinc/PU_1_2_2.dzn
- (-60.7%: 366.0s -> 143.7s) data/mznc2019/ptv/pax_model.mzn data/mznc2019/ptv/pax_period_20_1c.dzn
- (-73.9%: 77.2s -> 20.1s) data/mznc2019/fox-geese-corn/foxgeesecorn.mzn data/mznc2019/fox-geese-corn/foxgeesecorn_54.dzn
- (-79.7%: 65.0s -> 13.2s) data/mznc2020/is/model.mzn data/mznc2020/is/v1HjuSBQMb.dzn
- (-81.0%: 584.1s -> 110.8s) data/mznc2019/hrc/hrc.mzn data/mznc2019/hrc/exp1-1-5425.dzn

Objective Changes (>±10.0%):
=========================
- (-14.9%: MIN 47.00 -> 40.00) data/mznc2020/minimal-decision-sets/sparse_mds.mzn data/mznc2020/minimal-decision-sets/backache_train4.dzn
```