@glebbelov 2021-03-19
Can have acceptance testing for this so we can modify features safely. For example, providing 2 fake statistics and checking the number/kind of outputs. To facilitate checking, the output could be in YAML/TOML/JSON/etc