@Dekker1 2021-03-19
I suppose my idea was that the `PerformanceChanges` object would be used directly if you want to further analyse the results. It would be possible to use this in any new scripts. The output was really only meant for human consumption, but we can also add a YAML/JSON/TOML output mode to the command

My current implementation already finds and report conflicting solving status's, but I haven't done anything yet for conflicting optimal objectives, which I probably should.