@Dekker1 2021-03-19
@glebbelov I've added the finding of conflicting objectives as well (which already proved quite useful on my current benchmark.

I'll see if I have time this weekend to add the serialised output mode